requirements.txt
Django==4.2.7
djangorestframework==3.14.0
requests==2.31.0
beautifulsoup4==4.12.2
celery==5.3.4
redis==5.0.1
python-decouple==3.8
geopy==2.4.1
pgeocode==0.4.1
psycopg2-binary==2.9.9
dj-database-url==2.1.0
gunicorn==21.2.0
whitenoise==6.6.0